+++ common/src/main/java/Hongik_SafeMap_Server/constant/EnvironmentConstant.java
@@ -1,0 +1,5 @@
+package Hongik_SafeMap_Server.constant;
+
+public class EnvironmentConstant {
+    public static final String LOCAL_SERVER_URL = "http://localhost:8080";
+}
