+++ common/src/main/java/Hongik_SafeMap_Server/constant/RequestFormatConstant.java
@@ -1,0 +1,5 @@
+package Hongik_SafeMap_Server.constant;
+
+public class RequestFormatConstant {
+    public static final String PASSWORD_REGEX = "^(?=.*[!@#$%^&*(),.?\":{}|<>])[a-zA-Z0-9!@#$%^&*(),.?\":{}|<>]{8,25}$";
+}
