+++ common/src/main/java/Hongik_SafeMap_Server/execption/MemberException.java
@@ -1,0 +1,5 @@
+package Hongik_SafeMap_Server.execption;
+
+public class MemberException extends RuntimeException {
+    public MemberException(String message) { super(message); }
+}
