+++ common/src/main/java/Hongik_SafeMap_Server/util/EnvironmentUtil.java
@@ -1,0 +1,23 @@
+package Hongik_SafeMap_Server.util;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.core.env.Environment;
+import org.springframework.stereotype.Component;
+
+import java.util.stream.Stream;
+
+@Component
+@RequiredArgsConstructor
+public class EnvironmentUtil {
+
+    private final Environment environment;
+
+    public Stream<String> getActiveProfiles() { return Stream.of(environment.getActiveProfiles()); }
+
+    public String getCurrentProfile() {
+        return getActiveProfiles()
+                .filter(profile -> profile.equalsIgnoreCase("prod") || profile.equalsIgnoreCase("dev"))
+                .findFirst()
+                .orElse("local");
+    }
+}
