+++ core-api/src/main/java/Hongik_SafeMap_Server/domain/member/domain/vo/MemberStatus.java
@@ -1,0 +1,5 @@
+package Hongik_SafeMap_Server.domain.member.domain.vo;
+
+public enum MemberStatus {
+    일반, 관리자
+}
