+++ core-api/src/main/java/Hongik_SafeMap_Server/domain/member/dto/SignUpResponse.java
@@ -1,0 +1,5 @@
+package Hongik_SafeMap_Server.domain.member.dto;
+
+public record SignUpResponse(String name,
+                             String email,
+                             String phone) {}
