+++ core-api/src/main/resources/application.yml
