+++ core-api/src/test/java/Hongik_SafeMap/Hongik_SafeMap_Server/HongikSafeMapServerApplicationTests.java
