--- src/main/java/Hongik_SafeMap/Hongik_SafeMap_Server/HongikSafeMapServerApplication.java
@@ -1,13 +1,0 @@
-package Hongik_SafeMap.Hongik_SafeMap_Server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HongikSafeMapServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(HongikSafeMapServerApplication.class, args);
-	}
-
-}
