--- src/main/java/Hongik_SafeMap/Hongik_SafeMap_Server/domain/member/domain/vo/MemberStatus.java
@@ -1,5 +1,0 @@
-package Hongik_SafeMap.Hongik_SafeMap_Server.domain.member.domain.vo;
-
-public enum MemberStatus {
-    일반, 관리자
-}
