--- src/main/java/Hongik_SafeMap/Hongik_SafeMap_Server/domain/member/dto/SignUpResponse.java
@@ -1,5 +1,0 @@
-package Hongik_SafeMap.Hongik_SafeMap_Server.domain.member.dto;
-
-public record SignUpResponse(String name,
-                             String email,
-                             String phone) {}
