--- src/main/java/Hongik_SafeMap/Hongik_SafeMap_Server/global/constant/EnvironmentConstant.java
@@ -1,5 +1,0 @@
-package Hongik_SafeMap.Hongik_SafeMap_Server.global.constant;
-
-public class EnvironmentConstant {
-    public static final String LOCAL_SERVER_URL = "http://localhost:8080";
-}
