--- src/main/java/Hongik_SafeMap/Hongik_SafeMap_Server/global/constant/RequestFormatConstant.java
@@ -1,5 +1,0 @@
-package Hongik_SafeMap.Hongik_SafeMap_Server.global.constant;
-
-public class RequestFormatConstant {
-    public static final String PASSWORD_REGEX = "^(?=.*[!@#$%^&*(),.?\":{}|<>])[a-zA-Z0-9!@#$%^&*(),.?\":{}|<>]{8,25}$";
-}
