--- src/main/java/Hongik_SafeMap/Hongik_SafeMap_Server/global/exception/ErrorMessage.java
@@ -1,9 +1,0 @@
-package Hongik_SafeMap.Hongik_SafeMap_Server.global.exception;
-
-public class ErrorMessage {
-    // Member
-    public static final String EMAIL_DOES_NOT_EXIST = "존재하지 않는 이메일입니다.";
-    public static final String PASSWORD_INVALID_FORMAT = "비밀번호는 특수문자, 영문자, 숫자를 포함한 8자리 이상 문자열입니다.";
-    public static final String DUPLICATED_EMAIL = "이미 존재하는 이메일입니다.";
-    public static final String PASSWORD_IS_DIFFERENT_FROM_CHECK = "비밀번호와 비밀번호 확인 입력값이 다릅니다.";
-}
