--- src/main/java/Hongik_SafeMap/Hongik_SafeMap_Server/global/exception/MemberException.java
@@ -1,5 +1,0 @@
-package Hongik_SafeMap.Hongik_SafeMap_Server.global.exception;
-
-public class MemberException extends RuntimeException {
-    public MemberException(String message) { super(message); }
-}
